scripts/requirements.txt
pyyaml
requests
psutil
